文档上传、保存、提取

+++ phoibe-doc-cms-domain/src/main/java/me/phoibe/doc/cms/domain/po/PhoibeDocument.java
@@ -80,7 +80,7 @@ public class PhoibeDocument {
 
     private String combatTypeString;
 
-    private int docType;
+    private Short docType;
 
     public Long getId() {
         return id;
@@ -386,11 +386,11 @@ public class PhoibeDocument {
         this.combatTypeString = combatTypeString;
     }
 
-    public int getDocType() {
+    public Short getDocType() {
         return docType;
     }
 
-    public void setDocType(int docType) {
+    public void setDocType(Short docType) {
         this.docType = docType;
     }
 }
